+++ CHANGELOG.md
@@ -8,7 +8,9 @@
 
 - ```startLoadingScreen``` : Function to call the loading screen of the framework.
 
-- ```removeLoadingScreen``` : Function to remove the loading screen from the page.
+- ```removeLoadingScreen``` : Function to remove the loading screen from the page
+
+- ```addSceneFromJson ``` : Function to load an existing Json file containing a scene and use it.
 
 ### Changed
 - ```create_copy(name, scale, counter, timer)``` :  Add the counter attribut which is an ``Integer``. If not included, by default it will be 0 and each time the function is call for the same model, it will create a copy with incrementing index. If counter is provided at the call, it will use the integer provided to add to the name of the copy and check if it already exist. If not it will add the new copy to the scene, else it will return and do nothing.